refactor: migrate file tree indices to size_t

+++ include/tui.h
@@ -16,9 +16,9 @@ typedef struct {
 
     int rows;                   // Ncurses LINES value
     int cols;                   // Ncurses COLS value
-    int visible_entries_head;   // Index of the first visible entry (inclusive)
-    int visible_entries_tail;   // Index of next entry of the last visible entry (exclusive)
-    int selected_entry;         // Index of the currently selected entry
+    size_t visible_entries_head;   // Index of the first visible entry (inclusive)
+    size_t visible_entries_tail;   // Index of next entry of the last visible entry (exclusive)
+    size_t selected_entry;         // Index of the currently selected entry
 
     int col_offset;             // Horizontal scroll offset for the tree view
     WINDOW *tree_pad;           // Pad for drawing the file tree
